feat(table): table 모듈 분리
- 임시 처리 로직 있음
