+++ module-table/src/main/java/kitchenpos/table/domain/Empty.java
