+++ module-table/src/main/java/kitchenpos/table/exception/CannotChangeEmptyException.java
