+++ module-table/src/main/java/kitchenpos/table/exception/CannotUngroupException.java
