+++ module-table/src/main/java/kitchenpos/table/exception/InvalidNumberOfGuestsException.java
