+++ module-table/src/test/java/kitchenpos/TableApplicationTests.java
@@ -1,0 +1,11 @@
+package kitchenpos;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TableApplicationTests {
+    @Test
+    public void contextLoads() {
+    }
+}
