+++ module-table/src/test/java/kitchenpos/table/domain/EmptyTest.java
