+++ module-table/src/test/java/kitchenpos/table/domain/NumberOfGuestsTest.java
