+++ module-table/src/test/resources/application.properties
@@ -1,0 +1,8 @@
+logging.level.web=DEBUG;
+logging.level.sql=DEBUG;
+logging.level.org.hibernate.type.descriptor.sql=trace
+
+spring.jpa.hibernate.ddl-auto=create-drop
+
+spring.jpa.properties.hibernate.show_sql=true
+spring.jpa.properties.hibernate.format_sql=true
